fix: offset 계산을 long 으로 바꿔 정수 오버플로 차단
page 는 아래쪽만 1로 보정되고 위쪽 상한이 없다. 요청 파라미터를 손으로 고쳐
page=300000000 을 보내면 (page-1)*size 가 int 범위를 넘어 -1294967306 으로 감싼다.
Java 는 정수 오버플로에 예외를 던지지 않고, 음수 OFFSET 도 오류가 아니라
조용히 빈 결과가 되므로 원인 추적이 어렵다.

Service 가 목록 조회 전에 페이지를 보정하므로 지금 경로에서는 드러나지 않지만,
그 보호는 한 메서드의 호출 순서에 있고 이 객체는 결재함과 근태 목록의 조건 객체로
재사용될 예정이다. 오버플로 방지는 계산을 소유한 이 객체가 책임진다.

+++ src/main/java/com/flowmate/org/domain/EmployeeSearchCond.java
@@ -65,8 +65,15 @@ public class EmployeeSearchCond {
         return size;
     }
 
-    /** SQL 의 OFFSET 값 */
-    public int getOffset() {
-        return (page - 1) * size;
+    /**
+     * SQL 의 OFFSET 값.
+     *
+     * long 으로 계산하는 이유: page 는 위쪽 상한이 없다(요청 파라미터를 손으로 고치면
+     * 얼마든 커진다). int 로 곱하면 Java 는 예외 없이 음수로 감싸고, 그 값이
+     * OFFSET 으로 들어가도 오류가 나지 않아 조용히 빈 결과가 된다.
+     * 오버플로 방지를 이 객체가 책임진다 — 호출하는 Service 의 순서에 의존하지 않는다.
+     */
+    public long getOffset() {
+        return (long) (page - 1) * size;
     }
 }
